aevol_propagate will now create a last_gener file

+++ src/libaevol/ExpManager.h
@@ -171,7 +171,7 @@ class ExpManager : public Observer
                          bool share_phenotypic_target);
     void write_setup_files(void);
     void save(void) const;
-    void save_copy(char* dir, int32_t num_gener = 0) const;
+    void save_copy(char* dir, int64_t num_gener = 0) const;
     inline void load(int32_t first_gener,
         bool verbose = false, bool to_be_run = true);
     void load(const char* dir, int64_t t0,
